validation pulled to upper level
