+++ api/app/Models/BaseModel.php
@@ -1,12 +1,9 @@
 <?php namespace App\Models;
 
-use App\Exceptions\ValidationException;
-use App\Services\SpiraValidator;
 use Carbon\Carbon;
 use Bosnadev\Database\Traits\UuidTrait;
 use DateTime;
-use Illuminate\Support\MessageBag;
-use Illuminate\Validation\Factory as Validator;
+use Illuminate\Validation\Factory as ValidationFactory;
 
 abstract class BaseModel extends \Spira\Repository\Model\BaseModel
 {
@@ -14,28 +11,9 @@ abstract class BaseModel extends \Spira\Repository\Model\BaseModel
 
     public $incrementing = false;
 
-    public $exceptionOnError = true;
-
     /**
-     * @var Validator
+     * @return ValidationFactory
      */
-    protected $validator;
-
-    /**
-     * @var MessageBag|null
-     */
-    protected $errors;
-
-    protected $validationRules = [];
-
-    /**
-     * @return array
-     */
-    public function getValidationRules()
-    {
-        return $this->validationRules;
-    }
-
     protected function getValidator()
     {
         if (is_null($this->validator)) {
@@ -48,14 +26,6 @@ abstract class BaseModel extends \Spira\Repository\Model\BaseModel
     public static function getTableName()
     {
         return with(new static())->getTable();
-    }
-
-    /**
-     * @return MessageBag|null
-     */
-    public function getErrors()
-    {
-        return $this->errors;
     }
 
     /**
@@ -83,67 +53,6 @@ abstract class BaseModel extends \Spira\Repository\Model\BaseModel
             default:
                 return $value;
         }
-    }
-
-    /**
-     * @param array $options
-     * @return bool|null
-     * @throws \Exception
-     */
-    public function save(array $options = [])
-    {
-        $result = true;
-        if ($this->fireModelEvent('validating') !== false) {
-            $result = $this->validate();
-        }
-
-        if (!$result && $this->exceptionOnError) {
-            throw new ValidationException($this->getErrors());
-        }
-
-        if ($this->fireModelEvent('validated') === false) {
-            return false;
-        }
-
-        return parent::save($options);
-    }
-
-    public function validate()
-    {
-        /** @var SpiraValidator $validation */
-        $validation = $this->getValidator()->make($this->attributes, $this->getValidationRules());
-        $validation->setModel($this);
-        $this->errors = [];
-        if ($validation->fails()) {
-            $this->errors = $validation->messages();
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
-     * Register a validating model event with the dispatcher.
-     *
-     * @param  \Closure|string  $callback
-     * @param  int  $priority
-     * @return void
-     */
-    public static function validating($callback, $priority = 0)
-    {
-        static::registerModelEvent('validating', $callback, $priority);
-    }
-
-    /**
-     * Register a validated model event with the dispatcher.
-     *
-     * @param  \Closure|string  $callback
-     * @param  int  $priority
-     * @return void
-     */
-    public static function validated($callback, $priority = 0)
-    {
-        static::registerModelEvent('validated', $callback, $priority);
     }
 
     /**
