+++ api/app/Services/SpiraValidator.php
@@ -1,16 +1,10 @@
 <?php namespace App\Services;
 
-use App\Models\BaseModel;
-use App\ValueObjects\TypeAwareMessage;
-use Illuminate\Validation\Validator;
 use Rhumsaa\Uuid\Uuid;
+use Spira\Repository\Validation\Validator;
 
 class SpiraValidator extends Validator
 {
-    /**
-     * @var BaseModel
-     */
-    protected $model;
 
     public function validateFloat($attribute, $value, $parameters)
     {
@@ -42,14 +36,6 @@ class SpiraValidator extends Validator
     }
 
     /**
-     * @param BaseModel $model
-     */
-    public function setModel(BaseModel $model)
-    {
-        $this->model = $model;
-    }
-
-    /**
      * Register custom validation rule for countries.
      *
      * @param  string  $attribute
@@ -64,19 +50,5 @@ class SpiraValidator extends Validator
             ->toArray();
 
         return in_array($value, array_fetch($countries, 'country_code'));
-    }
-
-    /**
-     * Replace all error message place-holders with actual values.
-     *
-     * @param  string  $message
-     * @param  string  $attribute
-     * @param  string  $rule
-     * @param  array   $parameters
-     * @return string
-     */
-    protected function doReplacements($message, $attribute, $rule, $parameters)
-    {
-        return new TypeAwareMessage(parent::doReplacements($message, $attribute, $rule, $parameters), $rule);
     }
 }
